Remove prior endpoint
- Migrate prior endpoint tests into experiment
- Add get single experiment endpoint
