+++ src/ert_storage/endpoints/experiments.py
@@ -3,8 +3,20 @@ from fastapi import APIRouter, Depends, Body
 from sqlalchemy.orm.attributes import flag_modified
 from ert_storage.database import Session, get_db
 from ert_storage import database_schema as ds, json_schema as js
-from ert_storage.endpoints.priors import experiment_priors_to_dict
-from typing import Any, Mapping, Optional, List
+from ert_storage.json_schema.prior import (
+    PriorConst,
+    PriorTrig,
+    PriorNormal,
+    PriorLogNormal,
+    PriorErtTruncNormal,
+    PriorStdNormal,
+    PriorUniform,
+    PriorErtDUniform,
+    PriorLogUniform,
+    PriorErtErf,
+    PriorErtDErf,
+)
+from typing import Any, Mapping, List, Type
 
 
 router = APIRouter(tags=["experiment"])
@@ -17,6 +29,14 @@ def get_experiments(
 ) -> List[js.ExperimentOut]:
     experiments = db.query(ds.Experiment).all()
     return [_experiment_from_db(exp) for exp in experiments]
+
+
+@router.get("/experiments/{experiment_id}", response_model=js.ExperimentOut)
+def get_experiment_by_id(
+    *, db: Session = Depends(get_db), experiment_id: UUID
+) -> js.ExperimentOut:
+    experiment = db.query(ds.Experiment).filter_by(id=experiment_id).one()
+    return _experiment_from_db(experiment)
 
 
 @router.post("/experiments", response_model=js.ExperimentOut)
@@ -102,6 +122,35 @@ def delete_experiment(*, db: Session = Depends(get_db), experiment_id: UUID) -> 
     experiment = db.query(ds.Experiment).filter_by(id=experiment_id).one()
     db.delete(experiment)
     db.commit()
+
+
+PRIOR_FUNCTION_TO_PYDANTIC: Mapping[ds.PriorFunction, Type[js.Prior]] = {
+    ds.PriorFunction.const: PriorConst,
+    ds.PriorFunction.trig: PriorTrig,
+    ds.PriorFunction.normal: PriorNormal,
+    ds.PriorFunction.lognormal: PriorLogNormal,
+    ds.PriorFunction.ert_truncnormal: PriorErtTruncNormal,
+    ds.PriorFunction.stdnormal: PriorStdNormal,
+    ds.PriorFunction.uniform: PriorUniform,
+    ds.PriorFunction.ert_duniform: PriorErtDUniform,
+    ds.PriorFunction.loguniform: PriorLogUniform,
+    ds.PriorFunction.ert_erf: PriorErtErf,
+    ds.PriorFunction.ert_derf: PriorErtDErf,
+}
+
+
+def prior_to_dict(prior: ds.Prior) -> dict:
+    return (
+        PRIOR_FUNCTION_TO_PYDANTIC[prior.function]
+        .parse_obj(
+            {key: val for key, val in zip(prior.argument_names, prior.argument_values)}
+        )
+        .dict()
+    )
+
+
+def experiment_priors_to_dict(experiment: ds.Experiment) -> Mapping[str, dict]:
+    return {p.name: prior_to_dict(p) for p in experiment.priors}
 
 
 def _experiment_from_db(exp: ds.Experiment) -> js.ExperimentOut:
